Implemented the PeriodBasedReport base class solving #1174
